chore: move SortDirection to common package
as its not tracker specific
