+++ dhis-2/dhis-api/src/main/java/org/hisp/dhis/common/SortDirection.java
@@ -25,7 +25,7 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package org.hisp.dhis.webapi.controller.event.mapper;
+package org.hisp.dhis.common;
 
 import java.util.Arrays;
 import lombok.AllArgsConstructor;
