+++ dhis-2/dhis-api/src/main/java/org/hisp/dhis/webapi/controller/event/mapper/OrderParam.java
@@ -28,6 +28,7 @@
 package org.hisp.dhis.webapi.controller.event.mapper;
 
 import lombok.Value;
+import org.hisp.dhis.common.SortDirection;
 
 /**
  * Order parameter container to use within services.
